+++ src/main/java/com/empresa/AppTurismo/models/dao/IPaisDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.empresa.AppTurismo.models.entity.Paises;
 
-public interface IPaisDao extends MongoRepository<Paises, Integer>{
+public interface IPaisDao extends MongoRepository<Paises, Long>{
 
 }
