+++ src/main/java/com/empresa/AppTurismo/models/services/IPaisService.java
@@ -9,7 +9,7 @@ import com.empresa.AppTurismo.models.entity.Paises;
 
 public interface IPaisService {
 	public List<Paises> findAll();
-	public Optional<Paises> findById(int id);
+	public Optional<Paises> findById(Long id);
 	public Paises save(Paises pais);
-	public void delete(int id);
+	public void delete(Long id);
 }
